Add PlatformSettingsPropertyDrawerBase
