+++ Packages/UGF.EditorTools/Editor/IMGUI.PlatformSettings/PlatformSettingsPropertyDrawer.cs
@@ -1,21 +1,17 @@
 ﻿using System.Collections.Generic;
-using UGF.EditorTools.Editor.IMGUI.PropertyDrawers;
 using UGF.EditorTools.Runtime.IMGUI.PlatformSettings;
 using UnityEditor;
-using UnityEngine;
 
 namespace UGF.EditorTools.Editor.IMGUI.PlatformSettings
 {
     [CustomPropertyDrawer(typeof(PlatformSettings<>), true)]
-    public class PlatformSettingsPropertyDrawer : PropertyDrawerBase
+    internal class PlatformSettingsPropertyDrawer : PlatformSettingsPropertyDrawerBase
     {
-        protected PlatformSettingsDrawer Drawer { get; } = new PlatformSettingsDrawer();
-
         public PlatformSettingsPropertyDrawer()
         {
             var platforms = new List<BuildTargetGroup>();
 
-            PlatformSettingsEditorUtility.GetPlatformsAll(platforms);
+            PlatformSettingsEditorUtility.GetPlatformsAvailable(platforms);
 
             for (int i = 0; i < platforms.Count; i++)
             {
@@ -27,21 +23,7 @@ namespace UGF.EditorTools.Editor.IMGUI.PlatformSettings
             Drawer.SettingsCreated += OnDrawerSettingsCreated;
         }
 
-        protected override void OnDrawProperty(Rect position, SerializedProperty serializedProperty, GUIContent label)
-        {
-            SerializedProperty propertyGroups = serializedProperty.FindPropertyRelative("m_groups");
-
-            Drawer.DrawGUI(position, propertyGroups);
-        }
-
-        public override float GetPropertyHeight(SerializedProperty serializedProperty, GUIContent label)
-        {
-            SerializedProperty propertyGroups = serializedProperty.FindPropertyRelative("m_groups");
-
-            return Drawer.GetHeight(propertyGroups);
-        }
-
-        protected virtual void OnDrawerSettingsCreated(string name, SerializedProperty propertySettings)
+        protected override void OnDrawerSettingsCreated(string name, SerializedProperty propertySettings)
         {
             propertySettings.managedReferenceValue = null;
         }
