+++ Packages/UGF.EditorTools/Editor/IMGUI.PlatformSettings/PlatformSettingsPropertyDrawer.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 800045c9d55d4241be32f1fae8f79b5c
-timeCreated: 1601420960
+guid: 9082d85b85f94f2d80e010d585f5f3f0
+timeCreated: 1601488582
