+++ Packages/UGF.EditorTools/Editor/IMGUI.PlatformSettings/PlatformSettingsPropertyDrawerBase.cs
@@ -1,0 +1,27 @@
+﻿using UGF.EditorTools.Editor.IMGUI.PropertyDrawers;
+using UnityEditor;
+using UnityEngine;
+
+namespace UGF.EditorTools.Editor.IMGUI.PlatformSettings
+{
+    public abstract class PlatformSettingsPropertyDrawerBase : PropertyDrawerBase
+    {
+        protected PlatformSettingsDrawer Drawer { get; } = new PlatformSettingsDrawer();
+
+        protected override void OnDrawProperty(Rect position, SerializedProperty serializedProperty, GUIContent label)
+        {
+            SerializedProperty propertyGroups = serializedProperty.FindPropertyRelative("m_groups");
+
+            Drawer.DrawGUI(position, propertyGroups);
+        }
+
+        public override float GetPropertyHeight(SerializedProperty serializedProperty, GUIContent label)
+        {
+            SerializedProperty propertyGroups = serializedProperty.FindPropertyRelative("m_groups");
+
+            return Drawer.GetHeight(propertyGroups);
+        }
+
+        protected abstract void OnDrawerSettingsCreated(string name, SerializedProperty propertySettings);
+    }
+}
